+++ tests/testthat/test-create_rt_date.R
@@ -8,7 +8,6 @@ test_that("create_rt_data returns expected default values", {
   expect_equal(result$future_fixed, 1)
   expect_equal(result$fixed_from, 0)
   expect_equal(result$use_pop, 0)
-  expect_equal(result$stationary, 0)
   expect_equal(result$future_time, 0)
 })
 
@@ -17,7 +16,6 @@ test_that("create_rt_data handles NULL rt input correctly", {
 
   expect_equal(result$estimate_r, 0)
   expect_equal(result$future_fixed, 0)
-  expect_equal(result$stationary, 1)
 })
 
 test_that("create_rt_data handles custom rt_opts correctly", {
@@ -26,7 +24,6 @@ test_that("create_rt_data handles custom rt_opts correctly", {
     rw = 0,
     use_breakpoints = FALSE,
     future = "project",
-    gp_on = "R0",
     pop = Normal(mean = 1000000, sd = 100)
   )
 
@@ -34,7 +31,6 @@ test_that("create_rt_data handles custom rt_opts correctly", {
 
   expect_equal(result$estimate_r, 0)
   expect_equal(result$use_pop, 1)
-  expect_equal(result$stationary, 1)
   expect_equal(result$future_time, 7)
 })